update everything from the new 1.12 logic.

+++ src/main/java/com/gregtechceu/gtceu/utils/EntityDamageUtil.java
@@ -25,13 +25,13 @@ public class EntityDamageUtil {
     public static void applyTemperatureDamage(@NotNull LivingEntity entity, int temperature, float multiplier,
                                               int maximum) {
         if (temperature > 320) {
-            int damage = (int) ((multiplier * (temperature - 300)) / 50.0F);
+            float damage = multiplier * (temperature - 300) / 50.0F;
             if (maximum > 0) {
                 damage = Math.min(maximum, damage);
             }
             applyHeatDamage(entity, damage);
         } else if (temperature < 260) {
-            int damage = (int) ((multiplier * (273 - temperature)) / 25.0F);
+            float damage = multiplier * (273 - temperature) / 25.0F;
             if (maximum > 0) {
                 damage = Math.min(maximum, damage);
             }
@@ -43,7 +43,7 @@ public class EntityDamageUtil {
      * @param entity the entity to damage
      * @param damage the damage to apply
      */
-    public static void applyHeatDamage(@NotNull LivingEntity entity, int damage) {
+    public static void applyHeatDamage(@NotNull LivingEntity entity, float damage) {
         // do not attempt to damage by 0
         if (damage <= 0) return;
         if (!entity.isAlive()) return;
@@ -63,7 +63,7 @@ public class EntityDamageUtil {
      * @param entity the entity to damage
      * @param damage the damage to apply
      */
-    public static void applyFrostDamage(@NotNull LivingEntity entity, int damage) {
+    public static void applyFrostDamage(@NotNull LivingEntity entity, float damage) {
         // do not attempt to damage by 0
         if (damage <= 0) return;
         if (!entity.isAlive()) return;
@@ -91,7 +91,7 @@ public class EntityDamageUtil {
      * @param entity the entity to damage
      * @param damage the damage to apply
      */
-    public static void applyChemicalDamage(@NotNull LivingEntity entity, int damage) {
+    public static void applyChemicalDamage(@NotNull LivingEntity entity, float damage) {
         // do not attempt to damage by 0
         if (damage <= 0) return;
         if (!entity.isAlive()) return;
@@ -100,7 +100,7 @@ public class EntityDamageUtil {
             return;
 
         entity.hurt(GTDamageTypes.CHEMICAL.source(entity.level()), damage);
-        entity.addEffect(new MobEffectInstance(MobEffects.POISON, damage * 100, 1));
+        entity.addEffect(new MobEffectInstance(MobEffects.POISON, (int) (damage * 100), 1));
         // TODO advancements
         // if (entity instanceof ServerPlayer) AdvancementTriggers.CHEMICAL_DEATH.trigger((ServerPlayer) entity);
     }
